Remove now-unnecessary intermediary class
Error handling is roughly the same across the board with 1.21.5- dropped
