+++ internal/paper1_21_8/src/main/java/com/lishid/openinv/internal/paper1_21_8/player/OpenPlayer.java
@@ -1,20 +1,16 @@
 package com.lishid.openinv.internal.paper1_21_8.player;
 
-import com.lishid.openinv.internal.paper26_1.player.BaseOpenPlayer;
-import com.mojang.logging.LogUtils;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.ProblemReporter;
 import net.minecraft.world.level.storage.PlayerDataStorage;
-import net.minecraft.world.level.storage.TagValueOutput;
-import net.minecraft.world.level.storage.ValueOutput;
 import org.bukkit.craftbukkit.CraftServer;
 import org.jetbrains.annotations.NotNull;
-import org.slf4j.Logger;
 
 import java.nio.file.Path;
+import java.util.Optional;
 
-public class OpenPlayer extends BaseOpenPlayer {
+public class OpenPlayer extends com.lishid.openinv.internal.paper26_1.player.OpenPlayer {
 
   protected OpenPlayer(
       CraftServer server,
@@ -25,24 +21,12 @@ public class OpenPlayer extends BaseOpenPlayer {
   }
 
   @Override
-  protected void trySave(ServerPlayer player) {
-    Logger logger = LogUtils.getLogger();
-    // See net.minecraft.world.level.storage.PlayerDataStorage#save(EntityHuman)
-    try (ProblemReporter.ScopedCollector scopedCollector = new ProblemReporter.ScopedCollector(player.problemPath(), logger)) {
-      PlayerDataStorage worldNbtStorage = server.getServer().getPlayerList().playerIo;
-
-      CompoundTag oldData = isOnline()
-          ? null
-          : worldNbtStorage.load(player.getName().getString(), player.getStringUUID(), scopedCollector).orElse(null);
-      CompoundTag playerData = getWritableTag(oldData);
-
-      ValueOutput valueOutput = TagValueOutput.createWrappingWithContext(scopedCollector, player.registryAccess(), playerData);
-      player.saveWithoutId(valueOutput);
-
-      saveSafe(player, oldData, playerData, worldNbtStorage);
-    } catch (Exception e) {
-      LogUtils.getLogger().warn("Failed to save player data for {}: {}", player.getScoreboardName(), e);
-    }
+  protected Optional<CompoundTag> load(
+      PlayerDataStorage storage,
+      ServerPlayer player,
+      ProblemReporter.ScopedCollector collector
+  ) {
+    return storage.load(player.getName().getString(), player.getStringUUID(), collector);
   }
 
   @Override
@@ -52,6 +36,8 @@ public class OpenPlayer extends BaseOpenPlayer {
 
   @Override
   protected void remove(@NotNull CompoundTag tag, @NotNull String key) {
+    // Note that the method signature here is different, though the usage
+    // appears identical. This version does not return the removed tag.
     tag.remove(key);
   }
